Add Encoder interface
Adds the Encoder interface. Structured logging can be encoded and
written in many ways. Everything from XML to Protocol Buffer to lines of
ASCII text can be a log. Users of a Logger should not have to care about
the format of the final logging output. The Encoder allows us to
abstract away that dependency.

One note: it is still helpful in practice to be aware of the Encoder.
Making sure that whatever types are added to your Data implement
BinaryMarshaler and TextMarshaler interfaces is a good start.