Adds Filter type
The Filter type is the heart of this logging package. By sending the
Data structure through a series of filters, the Log method is able to
manipulate the Data before it is encoded. This could be adding
additional fields such as timestamps, application names, application or
API versions, counters, stack traces, or whatever else you may need to
log. This could also be removing fields such as stack traces, redundant
data, personally identifiable information, or anything else that you may
need to not log. This could also be used to stop the Data altogether by
setting the Data to nil. A nil Data value will be passed through
subsequent filters, but will not be encoded.