Adds Stack Filter
Adds a filter to add a stack trace when the log level exceeds the
set threshold. The stack trace will reflect the stack trace of the
filter, so it will always include the filter, the log, any wrappers
around the log, and then the caller of the log. This does limit the
effectiveness of the stack trace when used to log errors that do not
have context.