Correct usage of sync.Pool for reusing zipper
Gzip writers were being reset correctly in the free method, but the new
method was duplicating work. It should be assumed that the message
returned when the Pool calls its New function is the same as what is
returned with a Put method call. The New function was not returning an
initialized gzip.Writer, so that was being initialized for every call.